requirements.txt
streamlit
paramiko
pyyaml

